fix(shared): merge project and user opencode plugin detection
Stop dropping user-level plugin conflicts when an empty project config exists by collecting plugin entries from every supported OpenCode config path.

Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)
